ci(workflows): add ec2deploy.yaml for Docker image build and push
This commit introduces a new GitHub Actions workflow to automate the process of building and pushing Docker images to DockerHub. The workflow is triggered on pushes to the master branch and includes steps for checking out code, setting up Docker Buildx, logging into DockerHub, and building/pushing the Docker image.